+++ test/client/integration.test.ts
@@ -1,25 +1,26 @@
 import { arrayify } from "@ethersproject/bytes";
 import { assert } from "chai";
+import { providers } from "ethers";
 import { ethers } from "hardhat";
-import { buildStrategies } from "../../ts/client/contexts";
-import { DepositPool } from "../../ts/client/features/deposit";
+import { CoreAPI } from "../../ts/client/coreAPI";
 import {
     SimulatorPool,
     TransferPackingCommand
 } from "../../ts/client/features/transfer";
-import { SyncedPoint } from "../../ts/client/node";
 import { SyncerService } from "../../ts/client/services/syncer";
 import { PRODUCTION_PARAMS } from "../../ts/constants";
 import { deployAll } from "../../ts/deploy";
 import { deployKeyless } from "../../ts/deployment/deploy";
 import { Group, storageManagerFactory } from "../../ts/factory";
+import { Genesis } from "../../ts/genesis";
 import * as mcl from "../../ts/mcl";
 
 describe("Client Integration", function() {
     it("run", async function() {
         await mcl.init();
         const [signer] = await ethers.getSigners();
-        const genesisEth1Block = (await signer.provider?.getBlockNumber()) as number;
+        const provider = signer.provider as providers.Provider;
+        const genesisEth1Block = await provider.getBlockNumber();
         await deployKeyless(signer, false);
         const group = Group.new({ n: 32 });
         const storagePacker = await storageManagerFactory(group);
@@ -35,32 +36,26 @@ describe("Client Integration", function() {
 
         const simPool = new SimulatorPool(group, storagePacker.state);
         await simPool.setTokenID();
+        const genesis = await Genesis.fromContracts(
+            contracts,
+            parameters,
+            genesisEth1Block
+        );
+
+        const apiPacker = CoreAPI.new(storagePacker, genesis, provider, signer);
+        const apiSyncer = CoreAPI.new(storageSyncer, genesis, provider, signer);
 
         const packingCommand = new TransferPackingCommand(
-            parameters,
-            storagePacker,
+            apiPacker.parameters,
+            apiPacker.l2Storage,
             simPool,
-            contracts.rollup
+            apiPacker.rollup
         );
         for (let i = 0; i < 10; i++) {
             await packingCommand.packAndSubmit();
         }
 
-        const depositPool = new DepositPool(
-            parameters.MAX_DEPOSIT_SUBTREE_DEPTH
-        );
-        const strategies = buildStrategies(
-            contracts,
-            storageSyncer,
-            parameters,
-            depositPool
-        );
-        const syncpoint = new SyncedPoint(genesisEth1Block, 0);
-        const syncService = new SyncerService(
-            contracts.rollup,
-            syncpoint,
-            strategies
-        );
+        const syncService = new SyncerService(apiSyncer);
         await syncService.initialSync();
         assert.equal(storageSyncer.state.root, storagePacker.state.root);
     }).timeout(300000);
